backend/requirement.txt
# Python==3.10.12
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
idna==3.7
oauthlib==3.2.2
pip==22.0.2
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==59.6.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
django-cors-headers==4.3.1
